# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A Go service for investigating \"Deep inspection\" of VMs in VMware vSphere",
        "title": "VM Deep Inspection Demo API",
        "contact": {},
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/vms": {
            "get": {
                "description": "Get a list of all virtual machines with optional name filtering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vms"
                ],
                "summary": "List all virtual machines",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"web\"",
                        "description": "Filter VMs where name contains this string",
                        "name": "name_contains",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of virtual machines",
                        "schema": {
                            "$ref": "#/definitions/types.VMListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "vSphere connection unavailable",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vms/check": {
            "post": {
                "description": "Run validation checks on a VM snapshot. If check parameter is provided, runs that specific check. If omitted, runs all available checks.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vms"
                ],
                "summary": "Run validation checks on a VM snapshot",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"web-server-01\"",
                        "description": "Original VM name",
                        "name": "vms",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"inspection-snapshot\"",
                        "description": "Snapshot name",
                        "name": "snapshot",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"fstab\"",
                        "description": "Check type to run (fstab, disk-access). If omitted, runs all checks.",
                        "name": "check",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Check completed successfully",
                        "schema": {
                            "$ref": "#/definitions/types.CheckResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "VM or snapshot not found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vms/clone": {
            "post": {
                "description": "Create a linked clone from a VM snapshot for inspection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vms"
                ],
                "summary": "Create a clone from VM snapshot",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"web-server-01\"",
                        "description": "VM name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Clone request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CloneRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Clone created successfully",
                        "schema": {
                            "$ref": "#/definitions/types.CloneResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "VM or snapshot not found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vms/delete-clone": {
            "delete": {
                "description": "Delete a cloned VM created for inspection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vms"
                ],
                "summary": "Delete a cloned VM",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"web-server-01-clone-123\"",
                        "description": "Clone VM name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Clone deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Clone not found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vms/inspect-snapshot": {
            "post": {
                "description": "Run virt-inspector or virt-v2v-inspector on a VM snapshot using VDDK",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vms"
                ],
                "summary": "Inspect a VM snapshot directly",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"web-server-01\"",
                        "description": "Original VM name",
                        "name": "vms",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"inspection-snapshot\"",
                        "description": "Snapshot name",
                        "name": "snapshot",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"virt-inspector\"",
                        "description": "Inspector type: 'virt-inspector' (default) or 'virt-v2v-inspector'",
                        "name": "inspector",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inspection completed successfully",
                        "schema": {
                            "$ref": "#/definitions/types.VMInspectionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "VM or snapshot not found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vms/{name}": {
            "get": {
                "description": "Get detailed information about a specific virtual machine by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vms"
                ],
                "summary": "Get virtual machine details",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"web-server-01\"",
                        "description": "VM name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Virtual machine details",
                        "schema": {
                            "$ref": "#/definitions/types.VMDetailsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "VM not found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "vSphere connection unavailable",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vms/{name}/snapshot": {
            "post": {
                "description": "Create a snapshot for a specific virtual machine",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vms"
                ],
                "summary": "Create a VM snapshot",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"web-server-01\"",
                        "description": "VM name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Snapshot creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.SnapshotCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Snapshot created successfully",
                        "schema": {
                            "$ref": "#/definitions/types.SnapshotCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "VM not found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "vSphere connection unavailable",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vms/{name}/snapshot/{snapshotName}": {
            "delete": {
                "description": "Delete a snapshot from a specific virtual machine",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vms"
                ],
                "summary": "Delete a VM snapshot",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"web-server-01\"",
                        "description": "VM name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"snapshot-123\"",
                        "description": "Snapshot name",
                        "name": "snapshotName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Whether to consolidate the snapshot (default: false)",
                        "name": "consolidate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Snapshot deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/types.SnapshotDeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "VM or snapshot not found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "vSphere connection unavailable",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "types.CheckResponse": {
            "type": "object",
            "properties": {
                "all_valid": {
                    "type": "boolean",
                    "example": true
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.CheckResult"
                    }
                },
                "snapshot_name": {
                    "type": "string",
                    "example": "backup-snapshot"
                },
                "vm_name": {
                    "type": "string",
                    "example": "web-server-01"
                }
            }
        },
...
        "types.CloneRequest": {
            "type": "object",
            "required": [
                "snapshot_name"
            ],
            "properties": {
                "clone_name": {
                    "type": "string",
                    "example": "my-clone"
                },
                "snapshot_name": {
                    "type": "string",
                    "example": "backup-snapshot"
                }
            }
        },
        "types.CloneResponse": {
            "type": "object",
            "properties": {
                "clone_name": {
                    "type": "string",
                    "example": "vm-clone-123"
                },
                "message": {
                    "type": "string",
                    "example": "Clone created successfully"
                },
                "snapshot_name": {
                    "type": "string",
                    "example": "backup-snapshot"
                },
                "status": {
                    "type": "string",
                    "example": "completed"
                },
                "vm_name": {
                    "type": "string",
                    "example": "web-server-01"
                }
            }
        },
        "types.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "details": {
                    "type": "string",
                    "example": "VM ID is required"
                },
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                }
            }
        },
        "types.SnapshotCreateRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Backup before upgrade"
                },
                "memory": {
                    "type": "boolean",
                    "example": false
                },
                "name": {
                    "type": "string",
                    "example": "backup-snapshot"
                },
                "quiesce": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "types.SnapshotDeleteResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Snapshot deleted successfully"
                },
                "snapshot_name": {
                    "type": "string",
                    "example": "snapshot-123"
                },
                "status": {
                    "type": "string",
                    "example": "completed"
                },
                "vm_name": {
                    "type": "string",
                    "example": "web-server-01"
                }
            }
        },
...
        "types.VMDetailsResponse": {
            "type": "object",
            "properties": {
                "advanced": {
                    "$ref": "#/definitions/types.VMAdvancedInfo"
                },
                "current_snapshot": {
                    "type": "string",
                    "example": "snapshot-1"
                },
                "disks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.VMDisk"
                    }
                },
                "files": {
                    "$ref": "#/definitions/types.VMFileInfo"
                },
                "guest_info": {
                    "$ref": "#/definitions/types.VMGuestInfo"
                },
                "hardware": {
                    "$ref": "#/definitions/types.VMHardwareInfo"
                },
                "location": {
                    "$ref": "#/definitions/types.VMLocationInfo"
                },
                "metadata": {
                    "$ref": "#/definitions/types.VMMetadata"
                },
                "network_adapters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.VMNetworkAdapter"
                    }
                },
                "resources": {
                    "$ref": "#/definitions/types.VMResourceInfo"
                },
                "runtime": {
                    "$ref": "#/definitions/types.VMRuntimeInfo"
                },
                "snapshots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.VMSnapshot"
                    }
                },
                "storage": {
                    "$ref": "#/definitions/types.VMStorageSummary"
                },
                "tools": {
                    "$ref": "#/definitions/types.VMToolsInfo"
                },
                "vm": {
                    "$ref": "#/definitions/types.VM"
                }
            }
        },
...
        "types.VMInspectionResponse": {
            "type": "object",
            "properties": {
                "inspector_type": {
                    "type": "string",
                    "example": "virt-inspector"
                },
                "message": {
                    "type": "string",
                    "example": "Inspection completed successfully"
                },
                "snapshot_name": {
                    "type": "string",
                    "example": "backup-snapshot"
                },
                "status": {
                    "type": "string",
                    "example": "completed"
                },
                "virt_inspector": {},
                "virt_v2v": {},
                "vm_name": {
                    "type": "string",
                    "example": "web-server-01"
                }
            }
        },
        "types.VMListResponse": {
            "type": "object",
            "properties": {
                "datacenter": {
                    "type": "string",
                    "example": "Datacenter1"
                },
                "total": {
                    "type": "integer",
                    "example": 150
                },
                "vms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.VM"
                    }
                }
            }
        },
...
    }
}